+++ lib/src/internal/tabbed_view_provider.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
 import 'package:tabbed_view/src/tabbed_view.dart';
 import 'package:tabbed_view/src/tabbed_view_controller.dart';
 import 'package:tabbed_view/src/tabbed_view_menu_item.dart';
 
 /// Propagates parameters to internal widgets.
-class TabbedViewData {
-  TabbedViewData(
+@internal
+class TabbedViewProvider {
+  TabbedViewProvider(
       {required this.controller,
       this.contentBuilder,
       this.onTabClose,
@@ -16,9 +18,11 @@ class TabbedViewData {
       required this.selectToEnableButtons,
       this.closeButtonTooltip,
       this.tabsAreaButtonsBuilder,
-      this.draggableTabBuilder,
       required this.menuItems,
-      required this.menuItemsUpdater});
+      required this.menuItemsUpdater,
+      required this.onTabDrag,
+      required this.draggingTabIndex,
+      required this.onDraggableBuild});
 
   final TabbedViewController controller;
   final bool contentClip;
@@ -30,10 +34,15 @@ class TabbedViewData {
   final bool selectToEnableButtons;
   final String? closeButtonTooltip;
   final TabsAreaButtonsBuilder? tabsAreaButtonsBuilder;
-  final DraggableTabBuilder? draggableTabBuilder;
   final List<TabbedViewMenuItem> menuItems;
   final MenuItemsUpdater menuItemsUpdater;
+  final OnTabDrag onTabDrag;
+  final int? draggingTabIndex;
+  final OnDraggableBuild? onDraggableBuild;
 }
 
 /// Updater for menu items
 typedef MenuItemsUpdater = void Function(List<TabbedViewMenuItem>);
+
+/// Event that will be triggered when the tab drag start or end.
+typedef OnTabDrag = Function(int? tabIndex);