+++ lib/src/tab_data.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:tabbed_view/src/tab_button.dart';
 import 'package:tabbed_view/src/tab_leading_builder.dart';
+import 'package:tabbed_view/src/tabbed_view_controller.dart';
 
 /// The tab data.
 ///
@@ -27,7 +28,7 @@ import 'package:tabbed_view/src/tab_leading_builder.dart';
 /// See also:
 ///
 /// * [TabbedView.contentBuilder]
-class TabData extends ChangeNotifier {
+class TabData extends ChangeNotifier with TabIndex {
   TabData(
       {dynamic value,
       required String text,
@@ -35,6 +36,7 @@ class TabData extends ChangeNotifier {
       Widget? content,
       TabLeadingBuilder? leading,
       bool closable = true,
+      this.draggable = true,
       this.keepAlive = false})
       : _value = value,
         _text = text,
@@ -46,6 +48,8 @@ class TabData extends ChangeNotifier {
   /// Identifies the content of the tab in the tree
   final UniqueKey uniqueKey = UniqueKey();
   final bool keepAlive;
+
+  final bool draggable;
 
   dynamic _value;