+++ lib/src/theme/default_themes/minimalist_theme.dart
@@ -37,6 +37,7 @@ class MinimalistTheme {
       padding: EdgeInsets.fromLTRB(6, 3, 3, 3),
       paddingWithoutButton: EdgeInsets.fromLTRB(6, 3, 6, 3),
       decoration: BoxDecoration(color: colorSet[50]!),
+      draggingDecoration: BoxDecoration(color: colorSet[50]!),
       normalButtonColor: colorSet[900]!,
       hoverButtonColor: colorSet[900]!,
       disabledButtonColor: colorSet[400]!,